Story Game
In this game, the player takes on the role of Austin, a kid who loves programming, and must make choices that determine the outcome of the story. The game has a linear narrative with four different scenes, and the player must make the right choices to progress to the next scene. If the player completes the story, they unlock a secret ending that can be accessed by entering the phrase "easter egg" at any time during the game. The secret ending has two possible outcomes, depending on the player's choice.